adds game stats module/ moves all methods to
